Fix warnings

+++ AppDeferred/src/GBufferShader.hpp
@@ -48,7 +48,7 @@ class GBufferShader : public Shader {
             // Handle frame size changing
             Messenger::addReceiver<WindowFrameSizeMessage>(nullptr, [&](const Message &msg) {
                 const WindowFrameSizeMessage & m(static_cast<const WindowFrameSizeMessage &>(msg));
-                glm::vec2 frameSize = (static_cast<const WindowFrameSizeMessage &>(msg)).frameSize;
+                glm::ivec2 frameSize = (static_cast<const WindowFrameSizeMessage &>(msg)).frameSize;
                 auto gbuffer = MasterRenderer::getFBO("gbuffer");
                 gbuffer->textures[0]->width = gbuffer->textures[1]->width = gbuffer->textures[2]->width = frameSize.x;
                 gbuffer->textures[0]->height = gbuffer->textures[1]->height = gbuffer->textures[2]->height = frameSize.y;
